First version for pycharm
