+++ gradle.properties
@@ -10,8 +10,9 @@ ultimateEdition=Ultimate
 
 # IU, IC, PC, PY
 # IC for AndroidStudio
-ideType=IC
-ideVersion=232.8660.185
+ideType=PC
+#ideVersion=232.8660.185
+ideVersion=2023.2
 # ALL, NOJS
 buildType=NOJS
