+++ utbot-intellij-main/src/main/resources/META-INF/plugin.xml
@@ -1,0 +1,99 @@
+<!--suppress HttpUrlsUsage -->
+<idea-plugin>
+
+    <id>org.utbot.intellij.plugin.id</id>
+    <name>UnitTestBot</name>
+    <vendor>utbot.org</vendor>
+    <depends>com.intellij.modules.platform</depends>
+
+    <depends optional="true" config-file="META-INF/withJava.xml">com.intellij.modules.java</depends>
+    <depends optional="true" config-file="META-INF/withKotlin.xml">org.jetbrains.kotlin</depends>
+    <depends optional="true" config-file="META-INF/withPython.xml">com.intellij.modules.python</depends>
+    <depends optional="true" config-file="META-INF/withGo.xml">org.jetbrains.plugins.go</depends>
+    <depends optional="true" config-file="META-INF/withAndroid.xml">org.jetbrains.android</depends>
+    <depends optional="true" config-file="META-INF/withIdeaMaven.xml">org.jetbrains.idea.maven</depends>
+
+    <actions>
+        <action id="org.utbot.intellij.plugin.ui.actions.GenerateTestsAction"
+                class="org.utbot.intellij.plugin.ui.actions.GenerateTestsAction"
+                text="Generate Tests with UnitTestBot..."
+                description="Cover code with auto-generated tests">
+            <add-to-group group-id="ProjectViewPopupMenu"/>
+<!--            <add-to-group group-id="GenerateGroup" anchor="after" relative-to-action="JavaGenerateGroup1"/>-->
+            <keyboard-shortcut keymap="$default" first-keystroke="alt shift U"/>
+            <keyboard-shortcut keymap="$default" first-keystroke="alt U" second-keystroke="alt T"/>
+        </action>
+    </actions>
+
+    <extensions defaultExtensionNs="com.intellij">
+        <projectConfigurable dynamic="true" parentId="tools" instance="org.utbot.intellij.plugin.settings.Configurable"
+                                 id="org.utbot.intellij.plugin.settings.Configurable"
+                                 displayName="UnitTestBot"/>
+        <!--suppress PluginXmlValidity -->
+        <projectService serviceImplementation="org.utbot.intellij.plugin.settings.Settings" preload="true"/>
+        <registryKey defaultValue="false" description="Enable editing Kotlin test files" key="kotlin.ultra.light.classes.empty.text.range"/>
+<!--        <postStartupActivity implementation="org.utbot.intellij.plugin.ui.GotItTooltipActivity"/>-->
+        <projectModelModifier implementation="org.androidstudio.plugin.util.UtAndroidGradleJavaProjectModelModifierWrapper" order="first"/>
+<!--        <projectModelModifier implementation="org.utbot.intellij.plugin.util.UtMavenProjectModelModifier" order="first"/>-->
+<!--        <projectModelModifier implementation="org.utbot.intellij.plugin.util.UtIdeaProjectModelModifier" order="first"/>-->
+        <!--Documentation-->
+<!--        <customJavadocTagProvider implementation="org.utbot.intellij.plugin.javadoc.UtCustomJavaDocTagProvider"/>-->
+<!--        <lang.documentationProvider language="JAVA" order="first" implementationClass="org.utbot.intellij.plugin.javadoc.UtDocumentationProvider"/>-->
+    </extensions>
+
+    <!-- Minimum and maximum build of IDE compatible with the plugin -->
+    <description>
+        <![CDATA[
+        UnitTestBot is the tool for automated unit test generation and precise code analysis.
+        <br/><br/>
+        Discover UnitTestBot key features in our latest release:
+        <ul>
+            <li>generating ready-to-use test cases — with valid inputs, method bodies, assertions, and comments</li>
+            <li>maximizing <a href="https://en.wikipedia.org/wiki/Code_coverage">branch coverage</a> in <i>regression suite</i> while keeping the number of tests minimized</li>
+            <li>finding deeply hidden code defects and expressing them as tests</li>
+            <li>fine-tuned <a href="https://github.com/UnitTestBot/UTBotJava/wiki/Fine-tune-test-generation#mocking-settings">mocking</a>, including mocking static methods</li>
+            <li>representing all the <a href="https://github.com/UnitTestBot/UTBotJava/wiki/Get-use-of-test-results#read-test-descriptions">test descriptions</a> in a human-readable format</li>
+            <li>generating <a href="https://github.com/microsoft/sarif-tutorials/blob/main/README.md">SARIF</a> reports</li>
+            <li>innovative symbolic execution engine combined with a smart fuzzing platform</li>
+        </ul>
+        UnitTestBot supports the latest JDKs, JUnit 4, JUnit 5, TestNG, Mockito and is suitable for all popular operational systems.
+        <br/>
+        Try <a href="https://www.utbot.org/demo/?language=Java&source=public%20class%20Recursion%20%7B%0A%0A%20%20public%20int%20fib(int%20n)%20%7B%0A%20%20%20%20if%20(n%20%3C%200)%20throw%20new%20IllegalArgumentException()%3B%0A%20%20%20%20if%20(n%20%3D%3D%200)%20return%200%3B%0A%20%20%20%20if%20(n%20%3D%3D%201)%20return%201%3B%0A%20%20%20%20%0A%20%20%20%20return%20fib(n%20-%201)%20%2B%20fib(n%20-%202)%3B%0A%20%20%7D%0A%7D%0A">UnitTestBot online demo</a> to see how it generates tests for your code in real time.
+        <br/>
+        Contribute to UnitTestBot via <a href="https://github.com/UnitTestBot/UTBotJava/blob/main/CONTRIBUTING.md">GitHub</a>.
+        <br/>
+        Found a bug? File an <a href="https://github.com/UnitTestBot/UTBotJava/issues">issue</a>.
+        <br/>
+        Have an idea? Start a <a href="https://github.com/UnitTestBot/UTBotJava/discussions">discussion</a>.
+        ]]>
+    </description>
+    <change-notes>
+        <![CDATA[
+        Check out UnitTestBot new features and major improvements in the latest release:
+        <ul>
+            <li>Generating tests for Python and JavaScript</li>
+            <li>New fuzzing platform providing support for multiple languages</li>
+            <li>Improved test generation for Kotlin code</li>
+            <li>Multiprocess architecture based on the <a href="https://github.com/JetBrains/rd">Reactive Distributed communication framework</a> with advanced logging and debugging options</li>
+            <li>Symbolic execution engine with higher priority and fewer false UNSAT verdicts</li>
+            <li>UI/UX improvements for test sources root, <code>setting.properties</code> file, cancellation, timeout settings, notifications</li>
+            <li>Summaries with fine-grained settings</li>
+            <li>SARIF reports enabled by default, displaying sandbox-related and timeout test failures, properly addressing source files</li>
+            <li>Improved monitoring visualized with Grafana</li>
+            <li>Test generation bug fixes</li>
+            <li>Detailed documentation on UnitTestBot components including UnitTestBot architecture overview</li>
+        </ul>
+    ]]>
+    </change-notes>
+<!--    <extensions defaultExtensionNs="com.intellij">-->
+<!--        <globalInspection language="JAVA"-->
+<!--                          displayName="Errors detected by UnitTestBot"-->
+<!--                          groupPath="Java"-->
+<!--                          groupBundle="messages.InspectionsBundle"-->
+<!--                          groupKey="group.names.probable.bugs"-->
+<!--                          enabledByDefault="true"-->
+<!--                          level="ERROR"-->
+<!--                          implementationClass="org.utbot.intellij.plugin.inspection.UnitTestBotInspectionTool"/>-->
+<!--    </extensions>-->
+
+</idea-plugin>
