+++ utbot-intellij-main/src/main/resources/META-INF/withAndroid.xml
@@ -1,0 +1,3 @@
+<!--Optional dependency on org.jetbrains.android-->
+<extensions defaultExtensionNs="org.jetbrains.android">
+</extensions>
