+++ utbot-intellij-main/src/main/resources/META-INF/withGo.xml
@@ -1,0 +1,3 @@
+<!--Optional dependency on org.jetbrains.plugins.go-->
+<extensions defaultExtensionNs="org.jetbrains.plugins.go">
+</extensions>
