+++ utbot-intellij-main/src/main/resources/META-INF/withIdeaMaven.xml
@@ -1,0 +1,3 @@
+<!--Optional dependency on org.jetbrains.idea.maven-->
+<extensions defaultExtensionNs="org.jetbrains.idea.maven">
+</extensions>
