+++ utbot-intellij-main/src/main/resources/META-INF/withJS.xml
@@ -1,0 +1,4 @@
+<!--Optional dependency on org.jetbrains.android-->
+<extensions defaultExtensionNs="com.intellij">
+    <depends>JavaScript</depends>
+</extensions>
