+++ utbot-intellij-main/src/main/resources/META-INF/withJava.xml
@@ -1,0 +1,3 @@
+<!--Optional dependency on org.jetbrains.android-->
+<extensions defaultExtensionNs="com.intellij.modules.java">
+</extensions>
