+++ utbot-intellij-main/src/main/resources/META-INF/withKotlin.xml
@@ -1,0 +1,3 @@
+<!--Optional dependency on org.jetbrains.android-->
+<extensions defaultExtensionNs="org.jetbrains.kotlin">
+</extensions>
