+++ utbot-intellij-main/src/main/resources/META-INF/withPython.xml
@@ -1,0 +1,4 @@
+<!--Optional dependency on com.intellij.modules.python-->
+<extensions defaultExtensionNs="com.intellij">
+    <depends>com.intellij.modules.python</depends>
+</extensions>
