+++ utbot-intellij-main/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+# suppress inspection "HttpUrlsUsage" for whole file
+backing.service.possibleEndpoints=http://utbot.org
+backing.service.urlPath=/utbot/errors
+backing.service.port=11000
+request.timeoutMillis=5000
