+++ utbot-intellij-main/src/main/resources/bundles/UtbotBundle.properties
@@ -1,0 +1,10 @@
+# {0} - Test report url prefix, {1] - suffix
+test.report.force.mock.warning=<b>Warning</b>: Some test cases were ignored, because no mocking framework is installed in the project.<br>\
+Better results could be achieved by <a href="{0}{1}">installing mocking framework</a>.
+test.report.force.static.mock.warning=<b>Warning</b>: Some test cases were ignored, because mockito-inline is not installed in the project.<br>\
+Better results could be achieved by <a href="{0}{1}">configuring mockito-inline</a>.
+test.report.test.framework.warning=<b>Warning</b>: There are several test frameworks in the project.\
+To select run configuration, please refer to the documentation depending on the project build system:\
+<a href="https://docs.gradle.org/current/userguide/java_testing.html#sec:configuring_java_integration_tests">Gradle</a>, \
+<a href="https://maven.apache.org/surefire/maven-surefire-plugin/examples/providers.html">Maven</a> \
+or <a href="https://www.jetbrains.com/help/idea/run-debug-configuration.html#compound-configs">Idea</a>.
