+++ utbot-intellij-main/src/main/resources/inspectionDescriptions/UnitTestBotInspectionTool.html
@@ -1,0 +1,11 @@
+<html lang="en">
+<body>
+<p>Reports unchecked exceptions detected by UnitTestBot.</p>
+<p>Example:</p>
+<pre>
+void foo(int a) {
+    return 1 / a; // throws ArithmeticException when `a == 0`
+}
+</pre>
+</body>
+</html>
