+++ utbot-intellij-main/src/main/resources/log4j2.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Configuration>
+    <Appenders>
+        <!--     Idea catches plugin stdout log and wraps it in its own format, so in IDE only message is logged-->
+        <Console name="IdeaAppender" target="SYSTEM_OUT">
+            <PatternLayout pattern="| UtBot - %-25c{1} | %msg%n"/>
+        </Console>
+        <!--        When working as separate process - temporary log4j2.xml would be created, in which          -->
+        <!--        substring `ref="IdeaAppender"` will be replaced with `ref="EngineProcessAppender"`          -->
+        <!--        delete_this_comment_key identifies lines which will be deleted prior creating temp log4j2.xml  -->
+        <!--        You can set bufferIO=true and immediateFlush=false for better performance,                  -->
+        <!--        but possibly lose some logs on process death - log4j2 service might die unflushed           -->
+        <!--delete_this_comment_key
+        <RollingFile append="true" name="EngineProcessAppender" bufferedIO="false" immediateFlush="true" fileName="${env:UTBOT_LOG_DIR}utbot-engine-current.log" filePattern="${env:UTBOT_LOG_DIR}utbot-engine-%i.log" bufferSize="1048576">
+            <PatternLayout pattern="%d{HH:mm:ss.SSS} | %-5level | %-25c{1} | %msg%n"/>
+            <Policies>
+                <SizeBasedTriggeringPolicy size="20 MB"/>
+            </Policies>
+            <DefaultRolloverStrategy max="10"/>
+        </RollingFile>
+        delete_this_comment_key-->
+    </Appenders>
+    <Loggers>
+        <Logger name="org.utbot.intellij" level="info" additivity="false">
+            <AppenderRef ref="IdeaAppender"/>
+        </Logger>
+        <Logger name="org.utbot" level="info" additivity="false">
+            <AppenderRef ref="IdeaAppender"/>
+        </Logger>
+        <Root level="error">
+            <AppenderRef ref="IdeaAppender"/>
+        </Root>
+    </Loggers>
+</Configuration>
