+++ utbot-python/samples/my_cov.json
@@ -1,0 +1,1 @@
+{"covered":[{"start":4,"end":4},{"start":12,"end":15},{"start":17,"end":17}],"notCovered":[{"start":5,"end":10},{"start":16,"end":16}]}
