+++ utbot-ui-commons/src/main/kotlin/org/utbot/intellij/plugin/settings/Configurable.kt
